Patched loudly yelling cache module, when try/catch fails. Using manual file check instead.
git-svn-id: https://svn.midgard-project.org/midgard/trunk/midcom@16515 62f49d1c-1c16-0410-9156-d311ce72abac

+++ midcom.core/midcom/services/cache/backend/flatfile.php
@@ -113,12 +113,19 @@ class midcom_services_cache_backend_flatfile extends midcom_services_cache_backe
             return;
         }
         
-        try {
+        // Note: try below failed for some reason, checking manually if the file exists instead
+        if (!file_exists("{$this->_dirname}{$key}"))
+        {
+            return;
+        }
+        
+        try
+        {
             unlink("{$this->_dirname}{$key}");
         }
         catch(Exception $e)
         {
-            /* Do nothing, be quite */
+            /* Do nothing, be quiet */
         }
     }
